Refactor into modular components
